Refactored verify_password_reset_token method in AuthenticationService to use PasswordResetTokenReader instead of PasswordResetTokenUtil
